Fix random cube rotation corruption
I was not setting an initial value for the `cubeRotation` struct, and therefore it was not guaranteed to be zero when I created a new scene.  This would cause incorrect rotation values to be calculated during rotation calls.

Setting the values of the cube rotation struct to zero manually fixed it.