Refactored memory handling
Previously most of the allocations of consumer managed structs were done by calling a `*create()` function.  However, this caused memory leaks when dealing with lists, as we would be forced to create an instance, set it's values, then copy the memory into a spot in the packed list.

This would cause ambiguity of if the pointer being passed to the `list_append()` function should be free'ed or not, since the data is still being used.  If the data should be free you can't use the custom `*free()` for that struct because that would free pointers being referenced by the value inside the list.

Therefore the pattern was swapped around, such that instead there are `*init()` functions that take a newly allocated structure and initializes the internals of it.  This now means that we can create structs on the stack if we wanted, or with a custom allocator, and not worry.

It also means that we can now have our lists allocate space for a new item, then let us do the initialization inline.  This allows us to set the values in place in the packed list itself, instead of setting the values then copying to the final place in memory inside the list.