Triangle rendering now uses barycentric coordinates.
Triangle rendering now renders triangles pixels by pixels instead of scanline by scanline by computing barycentric coordinates.  This simplified a lot of the code to figure out interpolation values for normals and texturing.  This is probably slower than the scanline approach since it has to deal with pixels not in the triangle, but it opens up the possibility for parallelization.

Rendering vertices was removed as it showing vertices larger than one pixel becomes tricky when going pixel by pixel without a second pass.  A second pass can be done easily enough but I don't use it enough to bother.